requirements.txt
Flask==3.1.0
langchain==0.3.22
langchain_community==0.3.20
langchain_core==0.3.49
langchain_openai==0.3.11
langchain_pinecone==0.2.3
openai==1.70.0
pinecone==6.0.2
python-dotenv==1.1.0
setuptools==75.8.0
sentence-transformers==2.2.2
pypdf
pinecone[grpc]
langchain_experimental
-e .